#1351 moved smvinput related tests from testSmvFramework to testSmvInput
