+++ src/test/python/testSmvInput.py
@@ -17,6 +17,7 @@ import json
 from test_support.smvbasetest import SmvBaseTest
 from smv import *
 from smv.datasetrepo import DataSetRepo
+from smv.error import SmvDqmValidationError
 
 class SmvInputTest(SmvBaseTest):
     @classmethod
@@ -106,6 +107,79 @@ id:integer"""
         o = file(jsonfile, "w+")
         o.write(self._schema_json_str(schema))
         o.close()
+
+    def test_SmvCsvStringData(self):
+        fqn = "stage.modules.D1"
+        df = self.df(fqn)
+        expect = self.createDF("a:String;b:Integer", "x,10;y,1")
+        self.should_be_same(expect, df)
+
+    def test_SmvCsvStringData_with_error(self):
+        fqn = "stage.modules.D1WithError"
+        try:
+            df = self.df(fqn)
+        except SmvDqmValidationError as e:
+            self.assertEqual(e.dqmValidationResult["passed"], False)
+            self.assertEqual(e.dqmValidationResult["dqmStateSnapshot"]["totalRecords"], 3)
+            self.assertEqual(e.dqmValidationResult["dqmStateSnapshot"]["parseError"]["total"],2)
+
+    def test_SmvMultiCsvFiles(self):
+        self.createTempInputFile("multiCsvTest/f1", "col1\na\n")
+        self.createTempInputFile("multiCsvTest/f2", "col1\nb\n")
+        self.createTempInputFile("multiCsvTest.schema", "col1: String\n")
+
+        fqn = "stage.modules.MultiCsv"
+        df = self.df(fqn)
+        exp = self.createDF("col1: String", "a;b")
+        self.should_be_same(df, exp)
+
+    def test_SmvMultiCsvFilesWithUserSchema(self):
+        self.createTempInputFile("test3/f1", "col1\na\n")
+        self.createTempInputFile("test3/f2", "col1\nb\n")
+        self.createTempInputFile("test3.schema", "col1: String\n")
+
+        fqn = "stage.modules.MultiCsvWithUserSchema"
+        df = self.df(fqn)
+        exp = self.createDF("1loc: String", "a;b")
+        self.should_be_same(df, exp)
+
+    def test_SmvCsvFileWithUserSchema(self):
+        self.createTempInputFile("test3.csv", "col1\na\nb\n")
+        self.createTempInputFile("test3.schema", "col1: String\n")
+
+        fqn = "stage.modules.CsvFile"
+        df = self.df(fqn)
+        exp = self.createDF("1loc: String", "a;b")
+        self.should_be_same(df, exp)
+
+    def test_SmvSqlModule(self):
+        fqn = "stage.modules.SqlMod"
+        exp = self.createDF("a: String; b: String", "def,mno;ghi,jkl")
+        df = self.df(fqn)
+        self.should_be_same(df, exp)
+
+        # verify that the tables have been dropped
+        tablesDF = self.smvApp.sqlContext.tables()
+        tableNames = [r.tableName for r in tablesDF.collect()]
+        self.assertNotIn("a", tableNames)
+        self.assertNotIn("b", tableNames)
+
+    def test_SmvSqlCsvFile(self):
+        self.createTempInputFile("test3.csv", "a,b,c\na1,100,c1\na2,200,c2\n")
+        self.createTempInputFile("test3.schema", "a: String;b: Integer;c: String\n")
+
+        fqn = "stage.modules.SqlCsvFile"
+        df = self.df(fqn)
+        exp = self.createDF("a: String; b:Integer",
+             """a1,100;
+                a2,200""")
+        self.should_be_same(df, exp)
+
+        # verify that the table have been dropped
+        tablesDF = self.smvApp.sqlContext.tables()
+        tableNames = [r.tableName for r in tablesDF.collect()]
+        self.assertNotIn("a", tableNames)
+
 
     def test_SmvXmvFile_infer_schema(self):
         fqn = "stage.modules.Xml1"
