+++ src/test/python/testSmvInput/stage/modules.py
@@ -15,6 +15,76 @@ from smv import *
 from smv.functions import smvStrCat
 import pyspark.sql.functions as F
 
+class D1(SmvCsvStringData):
+    def schemaStr(self):
+        return "a:String;b:Integer"
+    def dataStr(self):
+        return "x,10;y,1"
+
+class D1WithError(SmvCsvStringData):
+    def schemaStr(self):
+        return "a:String;b:Integer"
+    def dataStr(self):
+        return "a1,10;a2,x;a3,2;a4,;a5,y"
+
+class MultiCsv(SmvMultiCsvFiles):
+    def dir(self):
+        return "multiCsvTest"
+
+class MultiCsvWithUserSchema(SmvMultiCsvFiles):
+    UserSchema = "1loc: String"
+
+    def dir(self):
+        return "test3"
+
+    def userSchema(self):
+        return self.UserSchema
+
+class CsvFile(SmvCsvFile):
+    UserSchema = "1loc: String"
+
+    def path(self):
+        return "test3.csv"
+
+    def userSchema(self):
+        return self.UserSchema
+
+class SqlCsvFile(SmvSqlCsvFile):
+    UserSchema = "a: String; b: Integer; c: String"
+
+    def path(self):
+        return "test3.csv"
+
+    def userSchema(self):
+        return self.UserSchema
+
+    def query(self):
+        return "select a, b from df"
+
+class SqlMod(SmvSqlModule):
+    def tables(self):
+        return {
+            "A": SqlInputA,
+            "B": SqlInputB
+        }
+
+    def query(self):
+        return "select a, b from A inner join B on A.ida = B.idb"
+
+class SqlInputA(SmvModule):
+    def requiresDS(self):
+        return []
+
+    def run(self, i):
+        return self.smvApp.createDF("ida: Integer; a: String", "1,def;2,ghi")
+
+class SqlInputB(SmvModule):
+    def requiresDS(self):
+        return []
+
+    def run(self, i):
+        return self.smvApp.createDF("idb: Integer; b: String", "2,jkl;1,mno")
+
 
 class Xml1(SmvXmlFile):
     def fullPath(self):
